Removed FIXED and PRECISION in CLASS(classy) documentation.
The documentation says the extra_args block is for "any param that CLASS understands,
for FIXED and PRECISION". But FIXED and PRECISION don't correspond to cobaya
parameter types (as in "fixed, sampled or derived") and I don't see
something called FIXED and PRECISION in CLASS either.
So removing this to avoid confusion with cobaya parameter types.